+++ packages/jsx/CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
-## Unreleased
+## 0.1.3 (2022-10-03)
 
 - Bump dependencies
 - Export webpack loader
 - Export docusaurus plugin
 - Support `react17` `defaultFramework`
 - Update types
+- Fix `@whyframe/core` peer dependency version
 
 ## 0.1.2 (2022-09-18)